refactor auth credentials provider
